Bestseller section completed.
Before I start/forget, as with the last commit, I am yet to
add the arrows that allow you scroll left or right on the
cards sections (both Vegetables and Bestsellers). These
qould be handled later. For now, I have a deadline to meet.

Okay. This commit:
-   Creates the best seller section using the reasoning
    behind nav bars to create its top section.
-   The other half is just a repetition of the cards
    section from the vegetables part of the product
    section, with a few tweaks for color differences.

I think that is pretty much it. On to the "Category"?
section? Lol. should be named something else really. Byeee